requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
psycopg==3.3.3
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
